Move supporting files so they're easier to import and remove dependencies that are already forked
