--- bin/README.md
@@ -1,1 +1,0 @@
-Executing `go install vole` from the root will place the vole binary here.
