--- src/github.com/vole/gouuid/Makefile
@@ -1,6 +1,0 @@
-include $(GOROOT)/src/Make.inc
-
-TARG=github.com/nu7hatch/uuid
-GOFILES=uuid.go
-
-include $(GOROOT)/src/Make.pkg
