--- src/github.com/vole/gravatar/README.md
@@ -1,19 +1,0 @@
-# gravatar
-
-gravatar is a library to ease access to
-[Gravatar API](http://gravatar.com/site/implement/).
-It implements a functionality to get avatars and decode JSON profiles.
-
-## Build status
-
-<a href="http://goci.me/project/github.com/ftrvxmtrx/gravatar">
-<img src="http://goci.me/project/image/github.com/ftrvxmtrx/gravatar" />
-</a>
-
-## Installation
-
-    $ go get github.com/ftrvxmtrx/gravatar
-
-## Documentation and examples
-
-[gravatar on go.pkgdoc.org](http://go.pkgdoc.org/github.com/ftrvxmtrx/gravatar)
