--- src/github.com/vole/gravatar/doc.go
@@ -1,3 +1,0 @@
-// Package gravatar implements a functionality to get avatars and decode JSON
-// profiles from Gravatar (http://gravatar.com).
-package gravatar
