--- src/github.com/vole/web/examples/hello.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-    "github.com/hoisie/web"
-)
-
-func hello(val string) string { return "hello " + val }
-
-func main() {
-    web.Get("/(.*)", hello)
-    web.Run("0.0.0.0:9999")
-}
